+++ src/main/java/io/dbtower/insight/internal/web/InsightController.java
@@ -7,7 +7,7 @@ import io.dbtower.operator.model.PartitionInfo;
 import io.dbtower.operator.model.LatencyPercentile;
 import io.dbtower.operator.model.IndexAdvice;
 import io.dbtower.operator.model.DeadlockEvent;
-import io.dbtower.insight.internal.PrometheusClient;
+import io.dbtower.insight.PrometheusClient;
 import io.dbtower.analysis.AiAnalyzer;
 import io.dbtower.analysis.QueryMasker;
 import io.dbtower.analysis.DeepAnalyzer;